refactor(aux): extract R4-C2 client cache/shutdown slice from auxiliary_client.py
Move the client-cache + close/shutdown block (source lines 6933-7306,
sha256 5e139636618beb09558c120877da698fe7fde1b545cd6b00e1de6504e64d2be2)
from agent/auxiliary_client.py into agent/aux_client_cache.py as part of
the god-file kill (#78635, epic #78647).

The godfile re-exports every moved name (functions, class, and module
state _client_cache/_client_cache_lock/_CLIENT_CACHE_MAX_SIZE) so all
existing import paths, unittest.mock.patch targets, and module-attribute
access (tests/agent/test_auxiliary_client.py, test_crossloop_client_cache.py,
test_async_httpx_del_neuter.py, etc.) resolve is-identical through the seam.

Stay-behind deps (resolve_provider_client, _normalize_aux_provider,
_normalize_main_runtime, _peek_pool_entry, _pool_runtime_api_key,
_pool_cache_hint, _MAIN_RUNTIME_FIELDS, _resolve_nous_runtime_api,
_create_openai_client, _to_async_client, utils.base_url_host_matches) are
lazy-imported inside function bodies, matching the file's cycle-avoidance
idiom.

Tests: tests/test_aux_cache_seam.py asserts identity for all 17 re-exported
names plus store/get/evict/shutdown behavior through the seam. Full direct
surface: 196 passed (188 pre-existing + 8 new), 0 failed.
